Move speculos.rs and emulator_http_transport to tests/test
