+++ cmd/crates/stellar-ledger/src/emulator_tests.rs
@@ -5,7 +5,6 @@ use soroban_env_host::xdr::{self, Operation, OperationBody, Uint256};
 use std::vec;
 
 use crate::hd_path::HdPath;
-use crate::test_fixtures::{emulator_http_transport::EmulatorHttpTransport, speculos::Speculos};
 use crate::{test_network_hash, Blob, Error, LedgerSigner};
 
 use std::sync::Arc;
@@ -21,6 +20,14 @@ use tokio::time::sleep;
 fn ledger(host_port: u16) -> LedgerSigner<impl Exchange> {
     LedgerSigner::new(get_http_transport("127.0.0.1", host_port).unwrap())
 }
+
+mod test_helpers {
+    pub mod test {
+        include!("../tests/test/mod.rs");
+    }
+}
+
+use test_helpers::test::{emulator_http_transport::EmulatorHttpTransport, speculos::Speculos};
 
 #[tokio::test]
 async fn test_get_public_key() {
